add webpack builder. apply es6 import/export import syntax. Re-write modules. Add few updates in service files
