+++ .gitignore
@@ -1,11 +1,17 @@
+# OS folders and files
+.DS_Store
+Thumbs.db
+
+#Code editors files
+.idea
+*.sublime.*
+.vscode
+
+# auxiliary folders and files
 /node_modules
+*.log*
 
-* text eol=lf
-
-*.png binary
-*.jpg binary
-*.jpeg binary
-*.webp binary
-*.woff binary
-*.woff2 binary
+#Webpack files (bundle)
+public/js/bundle.js
+public/js/bundle.js.map
