+++ public/index.html
@@ -44,10 +44,6 @@
 
       <!-- Контент -->
     </main>
-    <script src="../src/control.js"></script>
-    <script src="../src/filter.js"></script>
-    <script src="../src/board.js"></script>
-    <script src="../src/tasks.js"></script>
-    <script src="../src/main.js"></script>
+    <script src="./js/bundle.js"></script>
   </body>
 </html>
