+++ webpack.config.js
@@ -3,10 +3,16 @@ const path = require('path');
 
 module.exports = {
   mode: 'development',
-  entry: './src/main.js',
+  entry: './src/view/main.js',
   output: {
     path: path.join(__dirname, 'public/js'),
     filename: 'bundle.js',
   },
-  devtool: 'source-map'
+  devtool: 'source-map',
+  devServer: {
+    contentBase: path.join(__dirname, 'public'),
+    watchContentBase: true,
+    compress: true,
+    port: 8000
+  }
 };
